LPS-90813 Move upgrade logic to module
