+++ modules/apps/segments/segments-content-targeting-upgrade/src/main/java/com/liferay/segments/content/targeting/upgrade/internal/v1_0_0/util/OrganizationMemberRuleConverter.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.segments.internal.upgrade.v0_0_1.util;
+package com.liferay.segments.content.targeting.upgrade.internal.v1_0_0.util;
 
 import com.liferay.segments.criteria.Criteria;
 import com.liferay.segments.criteria.contributor.SegmentsCriteriaContributor;
-import com.liferay.segments.internal.criteria.contributor.UserSegmentsCriteriaContributor;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -25,23 +24,22 @@ import org.osgi.service.component.annotations.Reference;
  * @author Eduardo García
  */
 @Component(
-	immediate = true, property = "rule.converter.key=UserGroupMemberRule",
+	immediate = true, property = "rule.converter.key=OrganizationMemberRule",
 	service = RuleConverter.class
 )
-public class UserGroupMemberRuleConverter implements RuleConverter {
+public class OrganizationMemberRuleConverter implements RuleConverter {
 
 	@Override
 	public void convert(
 		long companyId, Criteria criteria, String typeSettings) {
 
-		_userSegmentsCriteriaContributor.contribute(
-			criteria, "(userGroupIds eq '" + typeSettings + "')",
+		_userOrganizationSegmentsCriteriaContributor.contribute(
+			criteria, "(organizationId eq '" + typeSettings + "')",
 			Criteria.Conjunction.AND);
 	}
 
-	@Reference(
-		target = "(segments.criteria.contributor.key=" + UserSegmentsCriteriaContributor.KEY + ")"
-	)
-	private SegmentsCriteriaContributor _userSegmentsCriteriaContributor;
+	@Reference(target = "(segments.criteria.contributor.key=user-organization)")
+	private SegmentsCriteriaContributor
+		_userOrganizationSegmentsCriteriaContributor;
 
 }
